Changed routing bugs in Navigation bar and TestHome (Login showing when refresh error)
