+++ src/Components/ShoppingCart/ShoppingCart.js
@@ -1,23 +1,39 @@
 import React, {Component} from 'react';
 import CartItem from "./CartItem";
+import PleaseLogin from "../Login/PleaseLogin";
 
 class ShoppingCart extends Component {
 
 
     render() {
 
+
         return (
-            <div className={"jumbotron bg-warning"}>
 
-                <h1>Shopping cart</h1>
+            (localStorage.getItem("isLoggedin") === "true") ? (
 
-                <h3>Username : {  window.atob(localStorage.getItem("token-username") )  }</h3>
 
-                <CartItem/>
-                <CartItem/>
-                <CartItem/>
+                <div className={"jumbotron bg-warning"}>
 
-            </div>
+                    <h1>Shopping cart</h1>
+
+                    <h3>Username : {window.atob(localStorage.getItem("token-username"))}</h3>
+
+                    <CartItem/>
+                    <CartItem/>
+                    <CartItem/>
+
+                </div>
+
+            ) : (
+
+
+                <PleaseLogin/>
+
+
+            )
+
+
         );
     }
 }
