+++ src/Components/WishList/WishList.js
@@ -1,28 +1,94 @@
 import React, {Component} from 'react';
 import Wishitem from "./Wishitem";
+import axios from 'axios';
+import PleaseLogin from "../Login/PleaseLogin";
+import {Link} from "react-router-dom";
 
 class WishList extends Component {
-    render() {
-        return (
-            < div className="Container mt-5">
-                <div className="card mb-5 ">
-                <h1>WishList</h1>
 
-                    <h3>Username : {window.atob(localStorage.getItem("token-username") )}</h3>
-
-                </div>
-
-                <div>
-                    <Wishitem/>
-                    <Wishitem/>
-                    <Wishitem/>
-                    <br/>
-                    <br/>
-                    <br/>
-                </div>
-            </div>
-        );
+    constructor(props) {
+        super(props);
+        this.state ={mern:[]};
+        this.state.customerId="";
     }
+
+    componentDidMount() {
+
+
+        let cus_id= window.atob(localStorage.getItem("token-username") );
+
+        axios.get('http://localhost:5000/mern/' + cus_id)
+            .then(response => {
+                this.setState({ mern: response.data });
+            })
+            .catch(function (error){
+                console.log(error);
+            })
+
+
+
+    }
+
+
+
+    render() {
+
+
+        return (
+
+            (localStorage.getItem("isLoggedin") === "true") ? (
+
+
+            < div class="Container mt-5 bg-info">
+                <div className="card mb-5 ">
+                    <h1 className="text-center">WishList</h1>
+                </div>
+                <div>
+                    {
+                        this.state.mern.map((value, index) => {
+                            return(
+                                <div>
+                                    <Wishitem
+                                        wi_pid={value.wish_productid}
+                                        wi_name={value.wish_name}
+                                        wi_price={value.wish_price}
+                                        wi_discount={value.wish_discount}
+                                    />
+                                </div>
+                            )
+                        })
+
+
+
+                    }
+                    <br/>
+                    <br/>
+                    <br/>
+                </div>
+
+                <Link className="btn btn-primary mb-3 "
+                      to={"/products_common"}>Go to products page</Link>
+
+            </div>
+
+             ):(
+
+
+                 <PleaseLogin/>
+
+
+             )
+
+
+
+        );
+
+
+
+
+    }
+
+
 }
 
 export default WishList;
